+++ index.js
@@ -1,10 +1,26 @@
-var debug = require('debug');
-var log = debug('selenium-drivex');
+var log = require('debug')('selenium-drivex');
 var async = require('async');
+function p(wd) {
+  //return a nodejs promise or webdriver promise
+  var promiz;
+  var wdPromiz = wd.promise.defer();
+  var fulfill = function (n) {
+    wdPromiz.fulfill(n);
+  };
+  var reject = function (err) {
+    wdPromiz.reject(err);
+  };
+  promiz = (global.Promise) ? new Promise(function (good, bad) {
+    fulfill = good;
+    reject = bad;
+  }) : wdPromiz.promise;
+  return {promise: promiz, fulfill, reject};
+};
 
 module.exports = function drivex(driver, wd) {
   var methods = {
     find: function (locator, el) {
+      log('finds:start %s', locator);
       return (el ? el : driver).findElement(locator);
     },
     /**
@@ -14,6 +30,7 @@ module.exports = function drivex(driver, wd) {
      * @returns {Promise} resolves to an array of WebElements or []
      */
     finds: function (locator, el) {
+      log('finds:start %s', locator);
       return (el ? el : driver).findElements(locator);
     },
     /**
@@ -23,7 +40,9 @@ module.exports = function drivex(driver, wd) {
      * @returns {Promise} resolves to true or false
      */
     present: function (locator, el) {
+      log('present:start %s', locator);
       return (el ? el : driver).findElements(locator).then(function (elts) {
+        log('present:success %s', locator);
         return elts.length > 0
       });
     },
@@ -34,7 +53,9 @@ module.exports = function drivex(driver, wd) {
      * @returns {Promise} resolves to true or rejected
      */
     visible: function (locator, el) {
+      log('visible:start %s', locator);
       return methods.find(locator, el).then(function (elt) {
+        log('visible:success %s', locator);
         return elt.isDisplayed();
       });
     },
@@ -46,12 +67,12 @@ module.exports = function drivex(driver, wd) {
      * @returns {Promise} resolves to true or throw error
      */
     waitForElement: function (locator, timeout, msg) {
-      log('waitForElement', locator);
+      log('waitForElement:start %s', locator);
       return driver.wait(wd.until.elementLocated(locator), timeout, msg).then(function () {
-        log('waitForElement: found', locator);
+        log('waitForElement:found %s', locator);
         return methods.find(locator);
       }, function (err) {
-        log('waitForElement', err);
+        log('waitForElement %s', err);
         log(err.stack);
         throw new Error(msg || '[drivex.waitForElement] Element not locatable for locator ' + showLocator(locator));
       });
@@ -64,7 +85,9 @@ module.exports = function drivex(driver, wd) {
      * @returns {WebElementPromise} resolves to WebElement or throw error
      */
     waitForElementPromise: function (locator, timeout, msg) {
+      log('waitForElementPromise:start %s', locator);
       function waitReturnElement() {
+        log('waitForElementPromise>waitReturnElement:start %s', locator);
         return methods.waitForElement(locator, timeout || 5000, msg);
       }
 
@@ -80,22 +103,23 @@ module.exports = function drivex(driver, wd) {
      */
     waitForElementVisible: function (locator, timeout, msg) {
 
-      log('waitForElementVisible', locator);
-      return driver.sleep(10).then(function () {
-          return driver.wait(function () {
-              return methods.present(locator);
-          }, timeout, msg).then(function () {
-              driver.wait(function () {
-                  return methods.visible(locator);
-              }, timeout, msg)
-          })
+      log('waitForElementVisible:start %s', locator);
+      // return driver.sleep(10).then(function () {
+      return driver.wait(function () {
+        log('waitForElementVisible:present %s', locator);
+        return methods.present(locator);
+      }, timeout, msg).then(function () {
+        return driver.wait(function () {
+          log('waitForElementVisible:visible %s', locator);
+          return methods.visible(locator);
+        }, timeout, msg)
+        // })
       }).then(function (isVisible) {
-        log('waitForElementVisible: ' + isVisible, locator);
+        log('waitForElementVisible:find %s, isVisible %s ', locator, isVisible);
         return methods.find(locator);
       }, function (err) {
-        log('waitForElementVisible', err);
-        log(err.stack);
-        throw new Error(msg || '[drivex.waitForElementVisible] Element not visible: ' + showLocator(locator));
+        log('waitForElementVisible:err %s', err);
+        throw new Error(msg || `[drivex.waitForElementVisible] Element not visible ${locator}`);
       });
     },
     /**
@@ -106,7 +130,9 @@ module.exports = function drivex(driver, wd) {
      * @returns {WebElementPromise} resolves to WebElement or throw error
      */
     waitForElementVisiblePromise: function (locator, timeout, msg) {
+      log('waitForElementVisiblePromise:start %s', locator);
       function waitVisibleReturnElement() {
+        log('waitForElementVisiblePromise>waitVisibleReturnElement:start %s', locator);
         return methods.waitForElementVisible(locator, timeout || 5000, msg);
       }
 
@@ -121,7 +147,8 @@ module.exports = function drivex(driver, wd) {
      * @returns {Promise} resolves to a WebeElement.click() (which resolves itself to a Promise
      */
     selectByOptionText: function (locator, optionText, parentWebElement) {
-      var d = wd.promise.defer();
+      var d = p(wd);
+      log('selectByOptionText:start, locator:%s, text: %s', locator, optionText);
       methods.find(locator, parentWebElement).then(function (selectEl) {
         selectEl.findElements(wd.By.css('option')).then(function (elts) {
           var current = 0;
@@ -160,6 +187,7 @@ module.exports = function drivex(driver, wd) {
      * @returns {Promise} resolves to a WebeElement.click() (which resolves itself to a Promise
      */
     selectByOptionValue: function (locator, optionValue, parentWebElement) {
+      log('selectByOptionValue:start, locator:%s, value: %s', locator, optionValue);
       return methods.find(locator, parentWebElement).then(function (selectEl) {
         return selectEl.findElement(wd.By.css('option[value=\'' + optionValue + '\']')).then(function (element) {
           return element.click();
@@ -179,22 +207,23 @@ module.exports = function drivex(driver, wd) {
     firstVisible: function (locatorObject, timeout) {
       var keyFound;
       var elementTests = [];
-      Object.keys(locatorObject).forEach(function(key) {
+      log('firstVisible:start, locatorObject:%o', locatorObject);
+      Object.keys(locatorObject).forEach(function (key) {
         var loc = locatorObject[key];
-        elementTests.push(function() {
-          return methods.waitForElementVisible(loc, 100).then(function() {
+        elementTests.push(function () {
+          return methods.waitForElementVisible(loc, 100).then(function () {
             keyFound = key;
             return true;
-          }, function(err) {
+          }, function (err) {
             return false;
           });
         });
       });
-      return driver.wait(function() {
+      return driver.wait(function () {
         var elementTest = elementTests.shift();
         elementTests.push(elementTest);
         return elementTest();
-      }, timeout || 5000).then(function() {
+      }, timeout || 5000).then(function () {
         return keyFound;
       });
     },
@@ -206,8 +235,9 @@ module.exports = function drivex(driver, wd) {
      * @returns {WebElementPromise} resolves to true or throw error
      */
     validateText: function (locator, parentWebElement, expectedText) {
-      var d = wd.promise.defer();
-      methods.find(locator, parentWebElement,expectedText).getText().then(function (actual){
+      var d = p(wd);
+      log('validateText:start, locator: %s, text: %s', locator, expectedText);
+      methods.find(locator, parentWebElement, expectedText).getText().then(function (actual) {
         log('validateText : actual : ' + actual + ' expected : ' + expectedText);
         if (actual === expectedText) {
           d.fulfill(true);
@@ -225,9 +255,10 @@ module.exports = function drivex(driver, wd) {
      * @param expected text
      * @returns {WebElementPromise} resolves to true or throw error
      */
-    validateAttributeValue: function (locator, parentWebElement,attribute, expectedText) {
-      var d = wd.promise.defer();
-      methods.find(locator, parentWebElement,expectedText).getAttribute(attribute).then(function (actual) {
+    validateAttributeValue: function (locator, parentWebElement, attribute, expectedText) {
+      var d = p(wd);
+      log('validateAttributeValue:start, locator: %s, text: %s', locator, expectedText);
+      methods.find(locator, parentWebElement, expectedText).getAttribute(attribute).then(function (actual) {
         log('validateAttributeValue : actual : ' + actual + ' expected : ' + expectedText);
         if (actual === expectedText) {
           d.fulfill(true);