+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # selenium-drivex changelog
 
+## Unreleased
+
+- disable promise manager for tests
+- resolve broken waitForElement promise chain
+- improve logging
+- add conditional use of node.js promise if present
+
 ## v2.0.0
 
 - don't use deprecated webdriver method